fix: remove app specific functions and consolidate misc #1
