--- lib/type/i-flow-purchase-event.ts
@@ -1,5 +1,0 @@
-export interface IFlowPurchaseEvent {
-  flowBlueprintId: number
-  sold: number
-  soldOut: boolean
-}
