+++ package.json
@@ -28,12 +28,14 @@
   "homepage": "https://github.com/samatechtw/onflow-ts#readme",
   "dependencies": {
     "@samatech/onflow-fcl-esm": "^0.5.4",
-    "elliptic": "^6.5.4"
+    "elliptic": "^6.5.4",
+    "sha3": "^2.1.4"
   },
   "devDependencies": {
     "@onflow/config": "^0.0.2",
     "@types/elliptic": "^6.4.14",
     "@types/jest": "^27.0.3",
+    "@types/node": "^17.0.13",
     "google-protobuf": "^3.19.1",
     "jest": "^27.4.5",
     "prettier": "^2.5.1",
